fixed DTO bug fixed bug where the staled movies where not removed

+++ app/src/main/java/pdm_1718i/yamda/data/server/DTOs.kt
@@ -38,7 +38,7 @@ data class MovieDetailResult(
         //val production_companies array[object]
         //val production_countries : array[object]
         val release_date: String,
-        val revenue : Int,
+        val revenue : Long,
         val runtime : Int?,
         //val spoken_languages array[object]
         val status : String,
